Retire the last two listing paths that bypassed FastObjectTable
`MaterializedView` held a plain `ListingTable`, so a view's own Parquet was
the one listing-backed table in the catalog that never pruned. It holds a
`FastObjectTable` now, like `ExternalTable` does. A view large enough to be
worth materializing is large enough to be worth pruning.

That leaves every listing-backed provider going through one type: `read_*`
table functions, `CREATE EXTERNAL TABLE`, and materialized views. The
remaining providers are not listing tables and should not be — Delta,
Iceberg, Icechunk and Lance each read their own metadata, and the system
schema is not files at all.

`prune_scan` in beacon-file-stats loses its last caller with that, and was
already a second implementation of `fast_object::prune::prune_plan`. Removed,
along with `split_at_scan` and `rebuild_over_scan` beside it. Its module is
now only the session handle, so it is named `handle` rather than
`scan_pruning`.

Its plan-shape tests were the only coverage of an invariant `prune_plan`
also depends on: descend a single-child chain to the scan, and fail open on
anything else, a union or a bare `EmptyExec` included. Rather than delete
them, `prune_plan`'s descent is extracted into the same two functions the
old path had, and the four tests move across to cover it.
