+++ beacon-db/beacon-file-stats/src/lib.rs
@@ -51,7 +51,7 @@ pub mod manifest;
 #[cfg(feature = "datafusion")]
 pub mod pruning;
 #[cfg(feature = "datafusion")]
-pub mod scan_pruning;
+pub mod handle;
 pub mod registry;
 pub mod scalar;
 pub mod segment;
@@ -64,7 +64,7 @@ pub use manifest::{Manifest, SegmentEntry};
 #[cfg(feature = "datafusion")]
 pub use pruning::{FileStatsPruningStatistics, prune_files};
 #[cfg(feature = "datafusion")]
-pub use scan_pruning::{FileStatsHandle, new_file_stats_handle, prune_scan, try_file_stats_from_session};
+pub use handle::{FileStatsHandle, new_file_stats_handle, try_file_stats_from_session};
 pub use registry::{AnalyzedFile, ReconcileReport, Registry};
 pub use scalar::{StatScalar, super_type};
 pub use segment::{ColumnStat, ColumnStats, SegmentBuilder, SegmentReader};
